requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
pandas==2.1.2
scikit-learn==1.3.2
Faker==24.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
